requirements.txt
# Django e componentes principais
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.1
django-cors-headers>=4.1.0

# Adicione estes pacotes se ainda não estiverem no seu requirements.txt
dj-database-url>=1.2.0
django-cors-headers>=4.1.0
django-redis>=5.3.0

# Banco de dados
psycopg2-binary>=2.9.6

# Processamento assíncrono
celery>=5.3.1
redis>=4.6.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.10.0
webdriver-manager>=3.8.6
lxml>=4.9.3

# Utilitários
python-dateutil>=2.8.2
python-dotenv>=1.0.0
Pillow>=10.0.0

# Desenvolvimento
pytest>=7.4.0
pytest-django>=4.5.2
black>=23.7.0
flake8>=6.1.0